Split app view
